+++ src/misc/changelog.ts
@@ -6,6 +6,11 @@ export const changelog = [
 		"Balance",
 		"• Decreased Avatar damage buff (50 -> 5 damage/level)",
 		"• Decreased charges on Goblin Land Mines (3 -> 2)",
+		"Bug Fixes",
+		"• Fixed copy errors",
+		"Refactors",
+		"• Jotye Farm ignores unit types and just works off buff",
+		"• Replaced all TriggerSleepActions with PolledWaits",
 	].join( "\n" ) },
 	{ title: "Version 8", content: [
 		"Balance",